requirements.txt
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.25.0
google-genai>=1.0.0
streamlit>=1.31.0
plotly>=5.18.0
